begin fixing tests
